api/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
langchain-openai>=0.0.2
langchain-groq>=0.0.1
langgraph>=0.0.20
langchain-community>=0.0.10
requests>=2.31.0
python-dotenv>=1.0.0
duckduckgo-search>=2.6.0
pydantic>=2.5.0